C++ Builder 2010 compilation.
git-svn-id: http://www.bci2000.org/svn/trunk@4202 7ef59cdf-0624-0410-a474-e34ce222aafa

+++ src/extlib/math/LinearPredictor.h
@@ -45,12 +45,12 @@ class LinearPredictor
   LinearPredictor& SetModelOrder( int inOrder )
     { mModelOrder = inOrder; return *this; }
   int ModelOrder() const
-    { return mModelOrder; }
+    { return static_cast<int>( mModelOrder ); }
 
   virtual void TransferFunction( const std::valarray<T>&, Ratpoly<T>& ) const = 0;
 
  protected:
-  int mModelOrder;
+  size_t mModelOrder;
 };
 
 #endif // LINEAR_PREDICTOR_H
